requirements.txt
altair==5.5.0
annotated-types==0.7.0
anyio==4.11.0
attrs==25.4.0
beautifulsoup4==4.14.2
blinker==1.9.0
brotli==1.2.0
cachetools==6.2.2
certifi==2025.11.12
cffi==2.0.0
chardet==3.0.4
charset-normalizer==3.4.4
click==8.1.8
colorama==0.4.6
cryptography==46.0.3
cssselect2==0.8.0
deep-translator==1.11.4
deprecation==2.1.0
diffusers==0.35.2
elevenlabs==2.24.0
extra-streamlit-components==0.1.71
ffmpeg==1.4
filelock==3.20.0
fonttools==4.60.1
fsspec==2025.10.0
gitdb==4.0.12
GitPython==3.1.45
greenlet==3.2.4
gTTS==2.5.4
h11==0.16.0
h2==3.2.0
hf-xet==1.2.0
hpack==3.0.0
hstspreload==2025.1.1
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.36.0
hyperframe==5.2.0
idna==2.10
importlib_metadata==8.7.0
Jinja2==3.1.6
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
MarkupSafe==3.0.3
mpmath==1.3.0
multidict==6.7.0
narwhals==2.12.0
networkx==3.5
numpy==2.3.5
packaging==25.0
pandas==2.3.3
pdfkit==1.0.0
pillow==12.0.0
postgrest==2.24.0
propcache==0.4.1
protobuf==6.33.1
pyarrow==21.0.0
pycparser==2.23
pydantic==2.12.4
pydantic_core==2.41.5
pydeck==0.9.1
pydub==0.25.1
pydyf==0.11.0
PyJWT==2.10.1
pyphen==0.17.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
PyYAML==6.0.3
realtime==2.24.0
referencing==0.37.0
regex==2025.11.3
requests==2.32.5
rfc3986==1.5.0
rpds-py==0.29.0
safetensors==0.7.0
setuptools==80.9.0
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.8
storage3==2.24.0
streamlit==1.51.0
StrEnum==0.4.15
supabase==2.24.0
supabase-auth==2.24.0
supabase-functions==2.24.0
sympy==1.14.0
tenacity==9.1.2
tinycss2==1.5.1
tinyhtml5==2.0.0
tokenizers==0.22.1
toml==0.10.2
torch==2.9.1
tornado==6.5.2
tqdm==4.67.1
transformers==4.57.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
upstash_redis==1.3.0
urllib3==2.5.0
watchdog==6.0.0
weasyprint==66.0
webencodings==0.5.1
websockets==15.0.1
yarl==1.22.0
zipp==3.23.0
zopfli==0.4.0
